Refactor: Split tests for each API into separate files
